Move to Sybil 3
The test_sybil.py change is to stop the example being picked up as a python codeblock that needs checking.
