+++ conftest.py
@@ -1,0 +1,32 @@
+from doctest import REPORT_NDIFF, ELLIPSIS
+
+from sybil import Sybil
+from sybil.parsers.doctest import DocTestParser, FIX_BYTE_UNICODE_REPR
+from sybil.parsers.codeblock import PythonCodeBlockParser
+from sybil.parsers.capture import parse_captures
+from sybil.parsers.skip import skip
+
+from testfixtures import TempDirectory
+from testfixtures.sybil import FileParser
+
+
+def sybil_setup(namespace):
+    # _tempdir is in case it's overwritten by a test.
+    namespace['tempdir'] = namespace['_tempdir'] = TempDirectory()
+
+
+def sybil_teardown(namespace):
+    namespace['_tempdir'].cleanup()
+
+
+pytest_collect_file = Sybil(
+    parsers=[
+        DocTestParser(optionflags=REPORT_NDIFF|ELLIPSIS|FIX_BYTE_UNICODE_REPR),
+        PythonCodeBlockParser(),
+        parse_captures,
+        FileParser('tempdir'),
+        skip,
+    ],
+    patterns=['*.txt', '*.py'],
+    setup=sybil_setup, teardown=sybil_teardown,
+).pytest()
